requirements.txt
Django==2.1.5
django-autoslug-iplweb==1.9.4
django-ckeditor==5.6.1
django-filter==2.0.0
django-js-asset==1.1.0
djangorestframework==3.9.0
Markdown==3.0.1
Pillow==5.4.0
pytz==2018.7
sorl-thumbnail==12.5.0
